requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
factory-boy==3.2.1
Faker==16.3.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing-inspect==0.8.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
